requirements.txt
# Dependências essenciais
Flask>=2.3.0,<3.0.0
python-decouple>=3.8
requests>=2.31.0
validate-docbr>=1.10.0

# Dependências opcionais - o app funciona sem elas
reportlab>=4.0.0,<5.0.0
Pillow>=10.0.0,<11.0.0
supabase>=2.3.0,<3.0.0
